{"schema":"decodex/commit/1","summary":"Extract overlay backend contract owner","authority":"manual"}
